requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
pydantic==2.12.5
httpx
python-dotenv
